+++ CdIts.Caffoa.Cli/Parser/ObjectParser.cs
@@ -93,7 +93,7 @@ public class ObjectParser
 
     private InterfaceModel ExtractInterface(IList<OpenApiSchema> schemaOneOf, OpenApiDiscriminator openApiDiscriminator)
     {
-        if (openApiDiscriminator?.PropertyName is null)
+        if (openApiDiscriminator.PropertyName is null)
             throw new CaffoaParserError("cannot create oneOf interface without discriminator property");
         var model = new InterfaceModel
         {