+++ demo/DemoV3/Model/LongRunningfunctionStatus.generated.cs
@@ -31,41 +31,11 @@ namespace DemoV3.Model {
         [JsonProperty("result")]
         public virtual AnyUser Result { get; set; }
 
-        public LongRunningfunctionStatus ToLongRunningfunctionStatus() {
-            var item = new LongRunningfunctionStatus();
-            item.UpdateWithLongRunningfunctionStatus(this);
-            return item;
-        }
-
-        /// <summary>
-        /// Replaces all fields with the data of the passed object
-        /// </summary>
-        public void UpdateWithLongRunningfunctionStatus(LongRunningfunctionStatus other) {
+        public LongRunningfunctionStatus(){}
+        public LongRunningfunctionStatus(LongRunningfunctionStatus other) {
             Status = other.Status;
             Result = other.Result?.ToAnyUser();
         }
-
-        /// <summary>
-        /// Merges all fields of LongRunningfunctionStatus that are present in the passed object with the current object.
-        /// If merge settings are not omitted, Arrays will be replaced and null value will replace existing values
-        /// </summary>
-        public void MergeWithLongRunningfunctionStatus(LongRunningfunctionStatus other, JsonMergeSettings mergeSettings = null) {
-            MergeWithLongRunningfunctionStatus(JObject.FromObject(other), mergeSettings);
-        }
-
-        /// <summary>
-        /// Merges all fields of LongRunningfunctionStatus that are present in the passed JToken with the current object.
-        /// If merge settings are not omitted, Arrays will be replaced and null value will replace existing values
-        /// </summary>
-        public void MergeWithLongRunningfunctionStatus(JToken other, JsonMergeSettings mergeSettings = null) {
-            mergeSettings ??= new JsonMergeSettings()
-            {
-                MergeArrayHandling = MergeArrayHandling.Replace,
-                MergeNullValueHandling = MergeNullValueHandling.Merge
-            };
-            var sourceObject = JObject.FromObject(this);
-            sourceObject.Merge(other, mergeSettings);
-            UpdateWithLongRunningfunctionStatus(sourceObject.ToObject<LongRunningfunctionStatus>());
-        }
+        public LongRunningfunctionStatus ToLongRunningfunctionStatus() => new LongRunningfunctionStatus(this);
     }
 }