+++ demo/DemoV3/Model/UserWithId.generated.cs
@@ -15,44 +15,13 @@ namespace DemoV3.Model {
         [JsonProperty("registrationDate")]
         public virtual DateTime RegistrationDate { get; set; }
 
-        public UserWithId ToUserWithId() {
-            var item = new UserWithId();
-            item.UpdateWithUserWithId(this);
-            return item;
-        }
-
-        public virtual AnyCompleteUser ToAnyCompleteUser() => ToUserWithId();
-
-        /// <summary>
-        /// Replaces all fields with the data of the passed object
-        /// </summary>
-        public void UpdateWithUserWithId(UserWithId other) {
-            UpdateWithUser(other);
+        public UserWithId(){}
+        public UserWithId(UserWithId other) : base(other) {
             Id = other.Id;
             RegistrationDate = other.RegistrationDate;
         }
-
-        /// <summary>
-        /// Merges all fields of UserWithId that are present in the passed object with the current object.
-        /// If merge settings are not omitted, Arrays will be replaced and null value will replace existing values
-        /// </summary>
-        public void MergeWithUserWithId(UserWithId other, JsonMergeSettings mergeSettings = null) {
-            MergeWithUserWithId(JObject.FromObject(other), mergeSettings);
-        }
-
-        /// <summary>
-        /// Merges all fields of UserWithId that are present in the passed JToken with the current object.
-        /// If merge settings are not omitted, Arrays will be replaced and null value will replace existing values
-        /// </summary>
-        public void MergeWithUserWithId(JToken other, JsonMergeSettings mergeSettings = null) {
-            mergeSettings ??= new JsonMergeSettings()
-            {
-                MergeArrayHandling = MergeArrayHandling.Replace,
-                MergeNullValueHandling = MergeNullValueHandling.Merge
-            };
-            var sourceObject = JObject.FromObject(this);
-            sourceObject.Merge(other, mergeSettings);
-            UpdateWithUserWithId(sourceObject.ToObject<UserWithId>());
-        }
+        public UserWithId(User other) : base(other) {}
+        public UserWithId ToUserWithId() => new UserWithId(this);
+        public virtual AnyCompleteUser ToAnyCompleteUser() => ToUserWithId();
     }
 }