+++ demo/DemoV3/Services/MaintainanceService.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Threading.Tasks;
-using Caffoa;
 using DemoV3.Model;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Azure.WebJobs.Extensions.DurableTask;